requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7
openpyxl>=3.1.0

# 桌面執行檔打包（僅 Windows 打包環境需要）
pywebview>=4.4.1
pyinstaller>=6.0
